Adjust errata decision wording
Issue #709
